REFACTORED: ViewController.updateCamera to use a pure swift method instead of calling flythrough_camera update method. ADDED: FlyCamera class

+++ VRTracer/flythrough_camera.c
@@ -16,7 +16,7 @@ void flythrough_camera_update(
     float eye_speed,
     float degrees_per_cursor_move,
     float max_pitch_rotation_degrees,
-    int delta_cursor_x, int delta_cursor_y,
+    float delta_cursor_x, float delta_cursor_y,
     int forward_held, int left_held, int backward_held, int right_held,
     int jump_held, int crouch_held,
     unsigned int flags)
